Reindent with tabs.  Fix stylistic changes
